pyproject.toml
[tool.poetry]
name = "budget-tracker"
version = "0.1.0"
description = ""
authors = ["Akbota <aqbotamaratqyzy@gmail.com>"]
readme = "README.md"
packages = [{include = "budget_tracker"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
